% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosketchr-package.R
\docType{package}
\name{biosketchr-package}
\alias{biosketchr}
\alias{biosketchr-package}
\title{biosketchr: biosketchr}
\description{
Provides the template and functions that make creating a Biosketch easier.
}
\author{
\strong{Maintainer}: Tyson S. Barrett \email{t.barrett@aggiemail.usu.edu} (0000-0002-2137-1391)

Authors:
\itemize{
  \item Daniel Anderson
}

}
\keyword{internal}
