% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nih_biosketch.R
\name{make_datetbl_nih}
\alias{make_datetbl_nih}
\title{Format positions data into a table for NIH}
\usage{
make_datetbl_nih(d, start_date, end_date, description, order = NULL,
  trailing_dashes = TRUE)
}
\arguments{
\item{d}{Dataset including the information}

\item{start_date}{Name of column denoting the year the position started or honor was awarded}

\item{end_date}{Name of column denoting the year the position ended}

\item{description}{Name of column describing the full description of the position/honor, including title, department, university, and location}

\item{order}{Optional integer argument listing the order of entries}

\item{trailing_dashes}{If end date is empty, should the start date have trailing dashes? Defaults to \code{TRUE}.}
}
\value{
LaTex \code{datetbl}
}
\description{
Transforms a standard R data frame into a LaTeX \code{datetbl}.
}
