% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsf_biosketch.R
\name{make_datetbl_nsf}
\alias{make_datetbl_nsf}
\title{Format positions data into a table for NSF}
\usage{
make_datetbl_nsf(d, start_date, end_date, position, location,
  order = NULL, trailing_dashes = TRUE)
}
\arguments{
\item{d}{Dataset including the information}

\item{start_date}{Name of column denoting the year the position started or honor was awarded}

\item{end_date}{Name of column denoting the year the position ended}

\item{position}{Name of column describing the name of the position}

\item{location}{Name of column denoting the location of position}

\item{order}{Optional integer argument listing the order of entries}

\item{trailing_dashes}{If end date is empty, should the start date have trailing dashes? Defaults to \code{TRUE}.}
}
\value{
LaTex \code{itemize}
}
\description{
Transforms a standard R data frame into a LaTeX \code{table}.
}
