% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nih_biosketch.R
\name{make_numbered_citations}
\alias{make_numbered_citations}
\title{Format Numbered Citations}
\usage{
make_numbered_citations(...)
}
\arguments{
\item{...}{keys from the`.bib` file in order of appearance in biosketch}
}
\value{
LaTex \code{enumerated list}
}
\description{
Formats a list of keys into a LaTeX \code{enumerated list} of references.
}
