% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nih_biosketch.R
\name{nih_biosketch}
\alias{nih_biosketch}
\title{NIH Biosketch format.}
\usage{
nih_biosketch(..., highlight = NULL, citation_package = "none")
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso",
"zenburn", and "haddock". Pass \code{NULL} to prevent syntax highlighting.}

\item{citation_package}{The LaTeX package to process citations, \code{natbib}
or \code{biblatex}. Use \code{none} if neither package is to be used.}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating NIH Biosketches. Does not use an official template but should approximate the format required by NIH. Is built on the template created by Paul Magwene found here: https://github.com/pmagwene/latex-nihbiosketch.
}
\details{
Possible arguments for the YAML header are:
\itemize{
  \item \code{name} Your Full Name
  \item \code{eracommons} Your ERA Commons ID
  \item \code{position} Description of your research position
  \item \code{education} Your education attainment (include any Bachelors, Masters, Doctorate, Postdoc). Several items are required for the education: 1) degree, 2) school, 3) date, and 4) field.
  \item \code{researchsupport} Both \code{ongoing} and \code{completed} research support received. Several items are required for the research support for both ongoing and completed: 1) grant, 2) PI, 3) dates, 4) title, 5) description, and 6) role.
  \item \code{output} Should be one of the formats from \code{biosketchr}
  }
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "nih_biosketch", package = "biosketchr")
}

}
