% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsf_biosketch.R
\name{nsf_biosketch}
\alias{nsf_biosketch}
\title{NSF Biosketch format.}
\usage{
nsf_biosketch(..., highlight = NULL, citation_package = "none")
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso",
"zenburn", and "haddock". Pass \code{NULL} to prevent syntax highlighting.}

\item{citation_package}{The LaTeX package to process citations, \code{natbib}
or \code{biblatex}. Use \code{none} if neither package is to be used.}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating NIH Biosketches. Does not use an official template but should approximate the format required by NIH. Is built on the template created by Paul Magwene found here: https://github.com/pmagwene/latex-nihbiosketch.
}
\details{
Possible arguments for the YAML header are:
\itemize{
  \item \code{name} Your Full Name (can include prefix such as Dr.)
  \item \code{department} Your current department
  \item \code{university} Your current university or company
  \item \code{email} Your current email
  \item \code{telephone} Your current telephone number
  \item \code{preparation} Both degrees earned and positions worked in relevant to proposal. Four items are required for the education: 1) location, 2) field, 3) degree, and 4) dates.
  \item \code{bibliography} The name of the bib file
  \item \code{bibliographystyle} The bib style
  \item \code{output} Should be one of the formats from \code{biosketchr}
  }
}
